+++ rest/src/main/java/org/seedstack/i18n/infrastructure/finders/jpa/TranslationJpaFinder.java
@@ -103,7 +103,7 @@ public class TranslationJpaFinder extends BaseJpaRangeFinder<TranslationRepresen
         List<Key> keys;
         // Get all the keys with their default translation
         if (range != null) {
-            keys = entityManager.createQuery(q).setFirstResult((int) range.getOffset()).setMaxResults(range.getSize()).getResultList();
+            keys = entityManager.createQuery(q).setFirstResult((int) range.getOffset()).setMaxResults((int)range.getSize()).getResultList();
         } else {
             keys = entityManager.createQuery(q).getResultList();
         }
